subscriber/requirements.txt
confluent-kafka
flask
sib_api_v3_sdk
jproperties
requests
python-dotenv
pybreaker